requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
pydantic
python-dotenv
httpx
streamlit
google-generativeai
elevenlabs
